refactor: auth-server 불필요한 코드 삭제
