--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/application/user/UserFacade.kt
@@ -1,16 +1,0 @@
-package org.clonecoder.auth.application.user
-
-import org.clonecoder.auth.domain.user.UserCommand
-import org.clonecoder.auth.domain.user.UserService
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-
-@Service
-class UserFacade(
-    private val userService: UserService
-) {
-    @Transactional
-    fun register(user: UserCommand.RegisterUser) {
-        userService.register(user)
-    }
-}
