--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/common/exception/DuplicatedEmailException.kt
@@ -1,6 +1,0 @@
-package org.clonecoder.auth.common.exception
-
-class DuplicatedEmailException(
-    override val message: String = "Duplicated Email"
-): RuntimeException(message) {
-}
