+++ platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/common/properties/JwtProperties.kt
@@ -6,7 +6,7 @@ import org.springframework.boot.context.properties.ConstructorBinding
 @ConstructorBinding
 @ConfigurationProperties(prefix = "security.jwt")
 data class JwtProperties(
-    val token: org.clonecoder.auth.common.properties.JwtProperties.TokenProperties
+    val token: TokenProperties
 ) {
 
     @ConstructorBinding
