+++ platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/config/SecurityConfig.kt
@@ -1,5 +1,6 @@
-package org.clonecoder.auth.common.security
+package org.clonecoder.auth.config
 
+import org.clonecoder.auth.common.security.SecurityRole
 import org.slf4j.LoggerFactory
 import org.springframework.context.annotation.Bean
 import org.springframework.http.HttpMethod
@@ -29,9 +30,9 @@ class SecurityConfig {
         http.csrf().disable()
             .httpBasic().disable()
             .authorizeRequests()
-            .antMatchers(HttpMethod.DELETE).hasRole(org.clonecoder.auth.common.security.SecurityRole.ADMIN.name)
-            .antMatchers(HttpMethod.PUT).hasRole(org.clonecoder.auth.common.security.SecurityRole.ADMIN.name)
-            .antMatchers(HttpMethod.PATCH).hasRole(org.clonecoder.auth.common.security.SecurityRole.ADMIN.name)
+            .antMatchers(HttpMethod.DELETE).hasRole(SecurityRole.ADMIN.name)
+            .antMatchers(HttpMethod.PUT).hasRole(SecurityRole.ADMIN.name)
+            .antMatchers(HttpMethod.PATCH).hasRole(SecurityRole.ADMIN.name)
             .antMatchers(HttpMethod.POST,"/users").permitAll()
             .anyRequest().authenticated()
