--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/domain/user/UserCommand.kt
@@ -1,14 +1,0 @@
-package org.clonecoder.auth.domain.user
-
-class UserCommand {
-
-    data class RegisterUser(
-        val email: String,
-        val password: String,
-    ) {
-        fun toEntity() = UserAccount(
-            email = email,
-            password = password
-        )
-    }
-}
