--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/domain/user/UserReader.kt
@@ -1,5 +1,0 @@
-package org.clonecoder.auth.domain.user
-
-interface UserReader {
-    fun existsByEmail(email: String): Boolean
-}
