--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/domain/user/UserService.kt
@@ -1,24 +1,0 @@
-package org.clonecoder.auth.domain.user
-
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-
-@Service
-class UserService(
-    private val userStore: UserStore,
-    private val userReader: UserReader,
-) {
-    @Transactional
-    fun register(user: UserCommand.RegisterUser) {
-        val userAccount = user.toEntity()
-        checkEmailDuplication(user.email)
-        userStore.register(userAccount)
-    }
-
-    private fun checkEmailDuplication(email: String) {
-        when (userReader.existsByEmail(email)) {
-            true -> throw org.clonecoder.auth.common.exception.DuplicatedEmailException()
-            false -> return
-        }
-    }
-}
