--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/domain/user/UserStore.kt
@@ -1,5 +1,0 @@
-package org.clonecoder.auth.domain.user
-
-interface UserStore {
-    fun register(userAccount: UserAccount)
-}
