--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/infrastructure/user/UserAccountRepository.kt
@@ -1,8 +1,0 @@
-package org.clonecoder.auth.infrastructure.user
-
-import org.clonecoder.auth.domain.user.UserAccount
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserAccountRepository: JpaRepository<UserAccount, Long> {
-    fun existsByEmail(email: String): Boolean
-}
