--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/infrastructure/user/UserStoreImpl.kt
@@ -1,14 +1,0 @@
-package org.clonecoder.auth.infrastructure.user
-
-import org.clonecoder.auth.domain.user.UserAccount
-import org.clonecoder.auth.domain.user.UserStore
-import org.springframework.stereotype.Component
-
-@Component
-class UserStoreImpl(
-    private val userAccountRepository: UserAccountRepository
-): UserStore {
-    override fun register(userAccount: UserAccount) {
-        userAccountRepository.save(userAccount)
-    }
-}
