--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/interfaces/user/dto/UserDto.kt
@@ -1,14 +1,0 @@
-package org.clonecoder.auth.interfaces.user.dto
-
-import org.clonecoder.auth.interfaces.validator.Password
-import javax.validation.constraints.Pattern
-
-class UserDto {
-
-    data class RegisterUserRequest(
-        @Pattern(regexp = "^(.+)@(.+)$")
-        val email: String,
-        @Password
-        val password: String,
-    )
-}
