--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/interfaces/user/mapper/UserDtoMapper.kt
@@ -1,15 +1,0 @@
-package org.clonecoder.auth.interfaces.user.mapper
-
-import org.clonecoder.auth.domain.user.UserCommand
-import org.clonecoder.auth.interfaces.user.dto.UserDto
-
-class UserDtoMapper {
-
-    companion object {
-        fun of(request: UserDto.RegisterUserRequest): UserCommand.RegisterUser =
-            UserCommand.RegisterUser(
-                email = request.email,
-                password = request.password,
-            )
-    }
-}
