--- platform-server/auth-server/src/main/kotlin/org/clonecoder/auth/interfaces/validator/Password.kt
@@ -1,15 +1,0 @@
-package org.clonecoder.auth.interfaces.validator
-
-import javax.validation.Constraint
-import javax.validation.Payload
-import kotlin.reflect.KClass
-
-@Target(AnnotationTarget.FIELD)
-@Retention
-@Constraint(validatedBy = [PasswordValidator::class])
-annotation class Password(
-    val message: String = "",
-    val groups: Array<KClass<*>> = [],
-    val payload: Array<KClass<out Payload>> = []
-)
-
