+++ platform-server/auth-server/src/main/resources/config/application.yml
@@ -28,11 +28,6 @@ spring:
   config:
     activate:
       on-profile: local
-  datasource:
-    url: jdbc:mysql://localhost:3306/expedia_authentication
-    username: root
-    password: root
-    driver-class-name: com.mysql.cj.jdbc.Driver
   jpa:
     properties:
       hibernate:
@@ -44,11 +39,6 @@ spring:
   config:
     activate:
       on-profile: test
-  datasource:
-    url: 'jdbc:h2:mem:authentication'
-    username: 'user'
-    password: ''
-    driver-class-name: org.h2.Driver
   jpa:
     properties:
       hibernate:
