--- platform-server/auth-server/src/test/kotlin/org/clonecoder/auth/acceptance/AcceptanceTest.kt
@@ -1,17 +1,0 @@
-package org.clonecoder.auth.acceptance
-
-import org.clonecoder.auth.support.DatabaseCleanup
-import io.kotest.core.spec.style.FeatureSpec
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.boot.web.server.LocalServerPort
-import org.springframework.test.context.ActiveProfiles
-import org.springframework.test.context.TestConstructor
-import org.springframework.test.context.TestConstructor.AutowireMode.ALL
-
-@ActiveProfiles("test")
-@TestConstructor(autowireMode = ALL)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class AcceptanceTest(
-    private val databaseCleanup: DatabaseCleanup,
-    @LocalServerPort private val port: Int
-): FeatureSpec()
