--- platform-server/auth-server/src/test/kotlin/org/clonecoder/auth/config/DomainTestConfig.kt
@@ -1,5 +1,0 @@
-package org.clonecoder.auth.config
-
-
-class DomainTestConfig {
-}
