--- platform-server/auth-server/src/test/kotlin/org/clonecoder/auth/payload/UserPayload.kt
@@ -1,21 +1,0 @@
-package org.clonecoder.auth.payload
-
-import org.clonecoder.auth.interfaces.user.dto.UserDto
-
-fun userRegisterSuccessPayloads(): UserDto.RegisterUserRequest {
-    val correctPassword = "123!abACC123"
-
-    return UserDto.RegisterUserRequest(
-        email = "designjava@naver.com",
-        password = correctPassword
-    )
-}
-
-fun userRegisterFailPayloads(): UserDto.RegisterUserRequest {
-    val invalidPassword = "1234"
-
-    return UserDto.RegisterUserRequest(
-        email = "designjava@naver.com",
-        password = invalidPassword
-    )
-}
